+++ db/access.php
@@ -64,13 +64,4 @@ $capabilities = [
             'manager' => CAP_ALLOW,
         ],
     ],
-    'mod/coassemble:viewanalytics' => [
-        'captype' => 'read',
-        'contextlevel' => CONTEXT_MODULE,
-        'archetypes' => [
-            'teacher' => CAP_ALLOW,
-            'editingteacher' => CAP_ALLOW,
-            'manager' => CAP_ALLOW,
-        ],
-    ],
 ];